# backend/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tela REST API",
    "version": "2026-09-28",
    "summary": "HTTP API of tela, a self-hostable markdown-native team wiki.",
...
    "contact": {
      "name": "tela",
      "url": "https://telawiki.com/support",
      "email": "tela@telawiki.com"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://github.com/zcag/tela/blob/main/LICENSE"
    }
  },
...
  "servers": [
    {
      "url": "https://telawiki.com",
      "description": "The instance serving this document. A self-hosted instance rewrites this to its own origin."
    }
  ],
  "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }],
...
  "paths": {
...
    "/api/spaces": {
      "get": {
        "tags": ["spaces"],
        "operationId": "listSpaces",
        "summary": "Spaces you can access",
        "description": "Resolved through direct membership and organization or group grants alike.",
        "responses": {
          "200": {
            "description": "Accessible spaces.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["spaces"],
                  "properties": {
                    "spaces": { "type": "array", "items": { "$ref": "#/components/schemas/SpaceListItem" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      },
      "post": {
        "tags": ["spaces"],
        "operationId": "createSpace",
        "summary": "Create a space",
        "description": "The creator becomes owner. Passing `org_id` makes the space organization-owned and grants the whole organization `editor`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "maxLength": 200 },
                  "slug": { "type": "string", "maxLength": 100, "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$", "description": "Derived from `name` when omitted." },
                  "org_id": { "type": ["integer", "null"], "format": "int64", "description": "You must be a member of the organization." }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/SpaceEnvelope" },
          "400": {
            "description": "`invalid_json`, `invalid_name`, or `invalid_slug`.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "402": { "$ref": "#/components/responses/QuotaExceeded" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": {
            "description": "`slug_conflict` — that slug is taken.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
    "/api/spaces/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/PathId" }],
      "get": {
        "tags": ["spaces"],
        "operationId": "getSpace",
        "summary": "Get a space",
        "responses": {
          "200": { "$ref": "#/components/responses/SpaceEnvelope" },
          "400": { "$ref": "#/components/responses/InvalidId" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      },
      "patch": {
        "tags": ["spaces"],
        "operationId": "updateSpace",
        "summary": "Rename, re-slug, describe, or publish a space",
        "description": "`editor+` for name, slug, and description. **Owner only** for `visibility` — flipping to `public` makes the whole space readable with no login (see the `public` tag).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                  "name": { "type": "string", "maxLength": 200 },
                  "slug": { "type": "string", "maxLength": 100, "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$" },
                  "visibility": { "$ref": "#/components/schemas/Visibility" },
                  "description": { "type": "string", "maxLength": 280, "description": "Empty string clears it." }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/SpaceEnvelope" },
          "400": {
            "description": "`invalid_json`, `no_fields`, `invalid_name`, `invalid_slug`, `invalid_visibility`, or `invalid_description`.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "402": {
            "description": "`upgrade_required` — publishing is not included in the owner's plan on this instance.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "`slug_conflict`.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      },
      "delete": {
        "tags": ["spaces"],
        "operationId": "deleteSpace",
        "summary": "Delete a space",
        "description": "Owner only.",
        "responses": {
          "204": { "description": "Deleted." },
          "400": { "$ref": "#/components/responses/InvalidId" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
...
    "/api/public/spaces/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/PathId" }],
      "get": {
        "tags": ["public"],
        "operationId": "getPublicSpace",
        "summary": "A published space",
        "description": "A private or non-existent space is one indistinguishable 404.",
        "security": [],
        "responses": {
          "200": {
            "description": "The space.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["space"],
                  "properties": { "space": { "$ref": "#/components/schemas/PublicSpace" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidId" },
          "404": { "$ref": "#/components/responses/NoSuchPublicSpace" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
    "/api/public/spaces/{id}/tree": {
      "parameters": [{ "$ref": "#/components/parameters/PathId" }],
      "get": {
        "tags": ["public"],
        "operationId": "getPublicSpaceTree",
        "summary": "The whole page tree of a published space",
        "description": "Structure plus card metadata (excerpt, reading time, cover, tags) — enough to render an index without fetching every page.",
        "security": [],
        "responses": {
          "200": {
            "description": "Pages.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["pages"],
                  "properties": {
                    "pages": { "type": "array", "items": { "$ref": "#/components/schemas/PublicTreeNode" } }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidId" },
          "404": { "$ref": "#/components/responses/NoSuchPublicSpace" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
    "/api/public/spaces/{id}/pages/{page_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/PathId" },
        { "name": "page_id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } }
      ],
      "get": {
        "tags": ["public"],
        "operationId": "getPublicSpacePage",
        "summary": "Read a page in a published space",
        "description": "No login. Publishing grants read only — there is no public write path anywhere under `/api/public/`.",
        "security": [],
        "responses": {
          "200": {
            "description": "The page.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["page"],
                  "properties": { "page": { "$ref": "#/components/schemas/PublicPage" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidId" },
          "404": { "$ref": "#/components/responses/NoSuchPublicSpace" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
...
    "/api/public/by-handle/{handle}/spaces/{slug}": {
      "parameters": [
        { "$ref": "#/components/parameters/Handle" },
        { "name": "slug", "in": "path", "required": true, "description": "The space slug within that account.", "schema": { "type": "string" } }
      ],
      "get": {
        "tags": ["public"],
        "operationId": "getPublicByHandleSpace",
        "summary": "A published space by owner handle and slug",
        "description": "The canonical addressing for public content: `/{handle}/{space-slug}` in the browser resolves through here.",
        "security": [],
        "responses": {
          "200": {
            "description": "The space.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["space"],
                  "properties": { "space": { "$ref": "#/components/schemas/PublicSpace" } }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NoSuchPublicSpace" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A personal access token: `Authorization: Bearer tela_pat_<43 chars>`. Scope `read` allows GET and HEAD only; `write` and `admin` allow any method. A token may be pinned to a single space."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tela_session",
        "description": "The browser session cookie set by `POST /api/auth/login`. HttpOnly, SameSite=Lax, 30 days."
      }
    },
    "parameters": {
      "PathId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64", "minimum": 1 }
      },
...
      "Handle": {
        "name": "handle",
        "in": "path",
        "required": true,
        "description": "A username or an organization slug — one namespace, case-insensitive.",
        "schema": { "type": "string" }
      }
    },
    "responses": {
...
      "SpaceEnvelope": {
        "description": "The space.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["space"],
              "properties": { "space": { "$ref": "#/components/schemas/Space" } }
            }
          }
        }
      },
...
      "InvalidId": {
        "description": "`invalid_id` — a path parameter was not a positive integer.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "`unauthorized` — no valid credential. A malformed bearer token never falls back to the session cookie.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "`forbidden` (no access, or the role is too low), `viewer_no_write`, `api_key_scope` (the token's scope forbids this method), or `api_key_space_scope` (the token is pinned to another space).",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "NotFound": {
        "description": "`not_found`.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NoSuchPublicSpace": {
        "description": "`not_found` — no such published space. Private and non-existent are indistinguishable.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "QuotaExceeded": {
        "description": "`quota_exceeded` — a plan limit (pages, storage) was reached.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "Internal": {
        "description": "`internal` — server error.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The error envelope for every JSON route. Branch on `code`, not on the message.",
        "required": ["error", "code"],
        "properties": {
          "error": { "type": "string", "description": "Human-readable message. Not stable — do not parse it." },
          "code": { "type": "string", "description": "Machine-readable code.", "examples": ["forbidden", "not_found", "api_key_scope"] },
          "public_path": { "type": "string", "description": "Only with `code: forbidden_public` — where to read this page without an account." }
        }
      },
      "Timestamp": {
        "type": "string",
        "description": "`YYYY-MM-DD HH:MM:SS` in UTC.",
        "examples": ["2026-09-09 14:03:21"]
      },
...
      "Visibility": {
        "type": "string",
        "enum": ["private", "public"],
        "description": "`public` makes the entire space readable with no account, at `/{handle}/{space-slug}`."
      },
...
      "Space": {
        "type": "object",
        "required": ["id", "name", "slug", "visibility", "description", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "visibility": { "$ref": "#/components/schemas/Visibility" },
          "description": { "type": "string" },
          "created_at": { "$ref": "#/components/schemas/Timestamp" },
          "updated_at": { "$ref": "#/components/schemas/Timestamp" },
          "my_role": { "type": "string", "enum": ["owner", "editor", "viewer"], "description": "Your effective role. Set on single-space responses, absent from listings." },
          "public_path": { "type": "string", "description": "Server-computed `/{handle}/{space-slug}`. Present only for a public space — never derive it client-side." }
        }
      },
      "SpaceListItem": {
        "allOf": [
          { "$ref": "#/components/schemas/Space" },
          {
            "type": "object",
            "required": ["member_count", "is_personal", "principals"],
            "properties": {
              "member_count": { "type": "integer" },
              "is_personal": { "type": "boolean" },
              "principals": {
                "type": "array",
                "description": "Organizations and groups the space is shared with.",
                "items": {
                  "type": "object",
                  "required": ["kind", "name"],
                  "properties": {
                    "kind": { "type": "string", "enum": ["org", "group"] },
                    "name": { "type": "string" }
                  }
                }
              },
              "owner_org": {
                "type": "object",
                "description": "Present when the space is organization-owned.",
                "properties": {
                  "id": { "type": "integer", "format": "int64" },
                  "name": { "type": "string" }
                }
              }
            }
          }
        ]
      },
      "PageProps": {
        "type": "object",
        "additionalProperties": true,
        "description": "Free-form page properties. Reserved keys are dropped. Two are load-bearing: `deck: true` makes the body a Slidev presentation, `sheet: true` makes it a spreadsheet — both are then stored verbatim, frontmatter included.",
        "examples": [{ "status": "draft", "owner": "platform-team" }]
      },
...
      "CardMeta": {
        "type": "object",
        "description": "Index-card metadata computed from a page, so a listing can be rendered without fetching bodies.",
        "properties": {
          "kind": { "type": "string", "enum": ["deck", "sheet"], "description": "Absent for an ordinary prose page." },
          "excerpt": { "type": "string" },
          "reading_minutes": { "type": "integer" },
          "cover": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } }
        }
      },
      "PublicSpace": {
        "type": "object",
        "required": ["id", "name", "slug", "visibility", "description"],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "visibility": { "$ref": "#/components/schemas/Visibility" },
          "description": { "type": "string" },
          "owner_handle": { "type": "string", "description": "The owning organization's slug when the space is organization-owned, otherwise the owner's username." }
        }
      },
...
      "PublicTreeNode": {
        "allOf": [
          {
            "type": "object",
            "required": ["id", "title", "parent_id", "position", "created_at", "updated_at"],
            "properties": {
              "id": { "type": "integer", "format": "int64" },
              "title": { "type": "string" },
              "parent_id": { "type": ["integer", "null"], "format": "int64" },
              "position": { "type": "integer", "format": "int64" },
              "created_at": { "$ref": "#/components/schemas/Timestamp" },
              "updated_at": { "$ref": "#/components/schemas/Timestamp" }
            }
          },
          { "$ref": "#/components/schemas/CardMeta" }
        ]
      },
      "PublicPage": {
        "type": "object",
        "required": ["id", "title", "body", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "body": { "type": "string", "description": "Canonical markdown." },
          "props": { "$ref": "#/components/schemas/PageProps" },
          "deck": {
            "type": "object",
            "description": "Present only on a presentation page.",
            "properties": {
              "present_path": { "type": "string" },
              "cover_path": { "type": "string" }
            }
          },
          "created_at": { "$ref": "#/components/schemas/Timestamp" },
          "updated_at": { "$ref": "#/components/schemas/Timestamp" },
          "author": { "type": "string" },
          "editor": { "type": "string" }
        }
      },
...
    }
  }
}